// onsite-program/submissions/week-4/day-2/Akanimoh-Johnson/typechain-types/SchoolManagementSystem.ts
export declare namespace SchoolManagementSystem {
  export type StudentStruct = {
    id: BigNumberish;
    name: string;
    age: BigNumberish;
    status: BigNumberish;
  };

  export type StudentStructOutput = [
    id: bigint,
    name: string,
    age: bigint,
    status: bigint
  ] & { id: bigint; name: string; age: bigint; status: bigint };
}

export interface SchoolManagementSystemInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "delete_student"
      | "get_all_students"
      | "get_student"
      | "register_student"
      | "students"
      | "update_status"
      | "update_student"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "delete_student",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "get_all_students",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "get_student",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "register_student",
    values: [string, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "students",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "update_status",
    values: [BigNumberish, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "update_student",
    values: [BigNumberish, string, BigNumberish]
  ): string;

  decodeFunctionResult(
    functionFragment: "delete_student",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "get_all_students",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "get_student",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "register_student",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "students", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "update_status",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "update_student",
    data: BytesLike
  ): Result;
}

export interface SchoolManagementSystem extends BaseContract {
  connect(runner?: ContractRunner | null): SchoolManagementSystem;
  waitForDeployment(): Promise<this>;

  interface: SchoolManagementSystemInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  delete_student: TypedContractMethod<
    [_index: BigNumberish],
    [void],
    "nonpayable"
  >;

  get_all_students: TypedContractMethod<
    [],
    [SchoolManagementSystem.StudentStructOutput[]],
    "view"
  >;

  get_student: TypedContractMethod<
    [_index: BigNumberish],
    [SchoolManagementSystem.StudentStructOutput],
    "view"
  >;

  register_student: TypedContractMethod<
    [_name: string, _age: BigNumberish],
    [void],
    "nonpayable"
  >;

  students: TypedContractMethod<
    [arg0: BigNumberish],
    [
      [bigint, string, bigint, bigint] & {
        id: bigint;
        name: string;
        age: bigint;
        status: bigint;
      }
    ],
    "view"
  >;

  update_status: TypedContractMethod<
    [_index: BigNumberish, _new_status: BigNumberish],
    [void],
    "nonpayable"
  >;

  update_student: TypedContractMethod<
    [_index: BigNumberish, _new_name: string, _new_age: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "delete_student"
  ): TypedContractMethod<[_index: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "get_all_students"
  ): TypedContractMethod<
    [],
    [SchoolManagementSystem.StudentStructOutput[]],
    "view"
  >;
  getFunction(
    nameOrSignature: "get_student"
  ): TypedContractMethod<
    [_index: BigNumberish],
    [SchoolManagementSystem.StudentStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "register_student"
  ): TypedContractMethod<
    [_name: string, _age: BigNumberish],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "students"
  ): TypedContractMethod<
    [arg0: BigNumberish],
    [
      [bigint, string, bigint, bigint] & {
        id: bigint;
        name: string;
        age: bigint;
        status: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "update_status"
  ): TypedContractMethod<
    [_index: BigNumberish, _new_status: BigNumberish],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "update_student"
  ): TypedContractMethod<
    [_index: BigNumberish, _new_name: string, _new_age: BigNumberish],
    [void],
    "nonpayable"
  >;

  filters: {};
}